+++ package.json
@@ -18,6 +18,7 @@
     "electron-is-dev": "^3.0.1",
     "jspdf": "^2.5.1",
     "kill-port": "^2.0.1",
+    "pdf-lib": "^1.17.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-icons": "^5.0.1",
